Added drop tables if app isn't in production mode

+++ database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php
@@ -21,6 +21,8 @@ return new class () extends Migration {
 
     public function down(): void
     {
-        Schema::dropIfExists('personal_access_tokens');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('personal_access_tokens');
+        }
     }
 };
