+++ database/migrations/2023_09_07_165319_create_listing_images_table.php
@@ -22,6 +22,8 @@ return new class () extends Migration {
 
     public function down(): void
     {
-        Schema::dropIfExists('listing_images');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('listing_images');
+        }
     }
 };
