+++ database/migrations/2023_10_15_203933_create_notifications_table.php
@@ -19,6 +19,8 @@ return new class () extends Migration {
 
     public function down(): void
     {
-        Schema::dropIfExists('notifications');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('notifications');
+        }
     }
 };
