+++ database/migrations/2023_12_26_174152_create_messages_table.php
@@ -32,6 +32,8 @@ return new class () extends Migration {
      */
     public function down(): void
     {
-        Schema::dropIfExists('messages');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('messages');
+        }
     }
 };
