+++ database/migrations/2023_12_26_174310_create_chat_user_table.php
@@ -31,6 +31,8 @@ return new class () extends Migration {
      */
     public function down(): void
     {
-        Schema::dropIfExists('chat_user');
+        if (!app()->isProduction()) {
+            Schema::dropIfExists('chat_user');
+        }
     }
 };
